+++ app/models/time_data.rb
@@ -1,7 +1,7 @@
 module TimeData
 
   def self.for(order_detail)
-    "#{self}::#{order_detail.product.type}".constantize.new(order_detail).time_data
+    "#{self}::#{order_detail.product.type}Presenter".constantize.new(order_detail)
   rescue NameError => e
     # noop, return nil
   end
