--- app/models/time_data/instrument.rb
@@ -1,15 +1,0 @@
-module TimeData
-
-  class Instrument
-
-    def initialize(order_detail)
-      @order_detail = order_detail
-    end
-
-    def time_data
-      @order_detail.reservation
-    end
-
-  end
-
-end
