+++ app/presenters/time_data/base_presenter.rb
@@ -1,0 +1,7 @@
+module TimeData
+
+  class BasePresenter < SimpleDelegator
+
+  end
+
+end
