+++ app/support/reservations/rendering.rb
@@ -9,25 +9,6 @@ module Reservations::Rendering
     actual_end_at || reserve_end_at
   end
 
-  def to_s
-    return super unless reserve_start_at && reserve_end_at
-    range = TimeRange.new(display_start_at, display_end_at).to_s
-    range += " (Canceled)" if canceled_at.present?
-    range
-  end
-
-  def reserve_to_s
-    TimeRange.new(reserve_start_at, reserve_end_at).to_s
-  end
-
-  def actuals_string
-    if actual_start_at.blank? && actual_end_at.blank?
-      "No actual times recorded"
-    else
-      TimeRange.new(actual_start_at, actual_end_at).to_s
-    end
-  end
-
   def as_calendar_object(options = {})
     calendar_object_default.merge(
       if order.present?
